IMS: Fix Ims UT issue after multiple DDS
Telephony suppose to close ImsManager instance
for the sub when ims is switched to other sub.
So that next time when ims switch back to this
sub it can create new UT interface for the sub.
To fix the issue close ImsManager instance when
ims switch to other sub.

Change-Id: I34dd956d27da2db493c95237f094f7526c63c33d
CRs-Fixed: 2007913