IMS: UT status from AP keep same as Modem after reset capability
A change is introduced in O with the assumption that UT will be
available only when IMS is registered. With the current
implementation, UT request will go over CS if IMS is not
registered as UT capability is resetting as well.

UT capability should not be reset (for IMS deregistration and for
IMS feature state not ready) and it should always depend on the
modem indication for UT capability.

Change-Id: I606522bcafa18280cd304c25fd819b1b042021c6
CRs-Fixed: 2090723