Fix modem DDS recommendation is ignored
When call is in ringing, modem might send DDS recommendation determined
by device's mode, but after the DDS recommendation comes, PhoneSwitcher
is able to overwrite it by default data SUB resulting in skipping DDS
switch.

To fix it, Telephony needs to block some events when call is ongoing and
consider ringing state for data during call, and the reason is that
Telephony is unable to differentiate device's mode in detail, and under
DSDS Telephony needs to keep data continuous when ringing.

Change-Id: Ia0fca5e35759d2187bf2cfe475a1b27ca08e15ec
CRs-Fixed: 3119697