Reset the waiting APNs after data call is successful
If the data call success and fail occurs alternatively, the accumulated
retry count can't be reset in the successful case leading to increase
the retry time for the next fail retry.

To fix it, need to reset the waiting APNs for reinitialzing the retry
manager once data call is successful.

Change-Id: Ib5c3c7d232bd7807438bc9ef80d7de192dc8d2ad
CRs-Fixed: 3135812