Fix SIM Contact not shown in hotswap scenario.
In the current issue, device is booted without SIM card
and SIM card is inserted while user is in Contacts app.
When user tries adding SIM contact, it is successfully
added in the SIM, but it is not shown in the SIM contact list.

This is because when user tries adding SIM contact even before
PBM module is UP, invalid adn capacity is returned to telephony
and SimPhoneBookRecordCache states are corrupted with invalid values,
and when valid adn capacity is received it is not handled properly.

Fix: Handle the phonebook capacity change event only if adn capacity
is valid.

Change-Id: Icfc21f4d77e977741b38005db926d45c81a4e44a
CRs-Fixed: 3423894