Add provision to customize auto data switch logic
In vanilla implementation, the mobile data switch of the non-DDS always
remains in switched-off state. Internet data will be allowed on it only
when it has been overridden by a policy, e.g., via auto data switch.

For evaluating auto data switch, the current code simply checks if the
internet data is allowed in the non-DDS. Move this check to a separate
method, so that it can be overridden in the derived classes where more
thorough checks can be performed.

Also, make some auto data switch related members as protected so that
those can be accessed by the derived classes.

Change-Id: I5744d555035eb1f7b27c526e227c35b2d18ef497
CRs-Fixed: 3470804