Fix DDS can't move to another emergency call SUB
When emergency call is ongoing on nDDS SUB, the SIM of DDS SUB is
removed, Telephony prevents evaluting data phone for smart temp DDS
switch, but right now smart temp DDS switch also depends on the
method of onEvalute, so that the block can be removed then.

To fix it, remove the blocks where the codes return directly during
a call and the unnecessary evaluting action for SUB ready.

Change-Id: I44f7083bdb83221b2e352fbabd5e6fbe8cc9200e
CRs-Fixed: 3491487