telephony: Disable 5G when wifi is connected
Keeping the device connected to 5G network when wifi is in use,
simply serves no purpose and just wastes power. We don't need to
keep 5G on unless we're on mobile data, and calls/SMS work just
fine over VoLTE/VoWiFi, so switch to 4G when wifi is connected
and back to 5G (if enabled by the user) when disconnected.

Change-Id: I0bdb7af59f9a6fda1d1d5ae3b294b6919a74547c