Fix adding SIM contact during hotswap scenario
In the current issue, SIM card is removed and inserted
while user is in Contacts app.When user tries adding SIM
contact,as soon as the SIM card is inserted, it is failing.

This is because at the time of adding SIM contact phonebook
capacity is invalid and the phonebook capacity changed
indication is ignored while update of a SIM contact is in
progress.

Fix: Handle phonebook capacity change when adn capacity is
invalid during SIM contact update.

Change-Id: Ic191bf53b7e6d0d018482f38177dffae3481d7bc
CRs-Fixed: 3524072